pyproject.toml
[tool.poetry]
name = "eurothermlib"
version = "0.1.0"
description = "Simple Python package for operating a Eurotherm controller ."
authors = ["Thomas Haeber <thomas.haeber@kit.edu>"]
readme = "README.md"
packages = [
    { include = "eurothermlib", from = "src" },
]

[tool.poetry.scripts]
eurotherm = 'eurothermlib.cli.cli:cli'

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
click = "^8.1.4"
numpy = "^1.25.1"
pint = ">=0.24"
rich = "^13.4.2"
grpcio = "^1.64.1"
omegaconf = "^2.3.0"
pydantic = "^2.7.4"
xarray = "^2024.6.0"
reactivex = "^4.0.4"
pymodbus = {extras = ["serial"], version = "^3.6.9"}
tenacity = "^8.4.2"
textual = "^0.71.0"
textual-plotext = "^0.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pylint = "^2.17.4"
jupyter = "^1.0.0"
black = "^23.7.0"
toolz = "^0.12.0"
pytest-watcher = "^0.4.2"
git-changelog = "^2.5.2"
pre-commit = "^3.7.1"
bump-my-version = "^0.21.1"
grpcio-tools = "^1.64.1"
grpc-stubs = "^1.53.0.5"
mypy-protobuf = "^3.6.0"
matplotlib = "^3.9.0"
textual-dev = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
skip-string-normalization = true
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  ^/foo.py    # exclude a file named foo.py in the root of the project
  | service_pb2_grpc.pyi  # exclude autogenerated Protocol Buffer files anywhere in the project
)
'''

[tool.bumpversion]
current_version = "0.1.0"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = true
ignore_missing_version = false
ignore_missing_files = false
tag = false
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = false
message = "Bump version: {current_version} → {new_version}"
commit_args = ""

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = '''^version\s*=\s*([\'\"\`])({current_version})\1'''
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "src/eurothermlib/__init__.py"

[tool.git-changelog]
convention = "angular"
in-place = false
output = "changelog.md"
provider = "github"
sections = "fix,feat,refactor,chore,style,build"
template = "angular"

[tool.ruff.per-file-ignores]
"service_pb2.py" = ["E501"]
"service_pb2.pyi" = ["E501"]
"service_pb2_grpc.py" = ["E501"]
"service_pb2_grpc.pyi" = ["E501"]
"utils.py" = ["F821"]
